+++ packages/adminui-vue/CHANGELOG.md
@@ -6,8 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- `UNavbarNotificationsButton`: added opening of unread notifications with the `messageType` === `system`
+ - `UNavbarNotificationsButton`: added opening of unread notifications with the `messageType` === `system`
   in the modal window for the user when logging into the system.
+ - `processing.saved`: pass method name `insert/update` as second argument into a `saved` hook on form
 
 ### Changed